requirements.txt
requests
python-dotenv
slack-sdk
slack-bolt
google-auth
google-auth-oauthlib
google-api-python-client
